Now we try to get rid of some boilerplate code, related to triggering re-fresh / re-render of the react component. -------------------------------------------------- Sunday, 15 September 2019 23.30.41:
Simplified the "cache injector" "logic".
--------------------------------------------------
Monday, 16 September 2019 0.23.53:

The "new" re-render trigger-er did not work.
Switched back to the old one. Need to look into
this later a bit more.

For now, I put this "project"/"refactoring" on hold
and proceed with implementing logic for User CRUD
functionality.

If the current architecture for re-rendering will become
exponentially annoyingly cumbersome then we will need
to come back to this possible solution to reduce
complexity - for the re-rendering and "cache injection"
"logic".

--------------------------------------------------