Proceeding with User CRUD client side logic. -------------------------------------------------- Monday, 16 September 2019 1.03.50:
Rearranging the files a bit.
Cleaning up the root directory.
Moving away .sh files, into a designated
sub-folder.
--------------------------------------------------
Monday, 16 September 2019 18.58.01:

Refactored the code a bit.
--------------------------------------------------
Monday, 16 September 2019 19.27.57:

In the middle of some refactoring for "more type safety",
we have now a few compiler errors.

It is about the declaration and implementation for the
"MainPage" react components.

That is, about their types. The two should be related.
--------------------------------------------------
Monday, 16 September 2019 23.18.36:

Code compiles but router complains about
some "unspecified path".

Need to fix this runtime error.
It is marked in the source as "todo-now".
--------------------------------------------------
Tuesday, 17 September 2019 2.54.04:

Everything works !

The new, more type safe "cache injector" / "wrapper"
and also the parameter passing from the URL into the component.

See the SumIntPage.

Next thing : onto the single User-CRUD.

--------------------------------------------------