Tuesday, 17 September 2019 12.59.24:
Trying to implement "login".

--------------------------------------------------
Tuesday, 17 September 2019 13.18.24 :

Waiting for some help on how to implement
"login" logic using the router.

--------------------------------------------------
Wednesday, 18 September 2019 0.43.35:

Cleaning up the code a bit.
--------------------------------------------------
Wednesday, 18 September 2019 1.41.23:

Added some "workaround text", as a "hack", the router
is not really refreshing when its parents state changes.

Which is weird/unexpected. ATM, I have no good ideas on
how to implement the login page functionality without
terrible hacks involving dreaded VARs crawling all over
my skin... :) ... LOL ... I mean, in the code... which
after coding 5 million hours... starts to feel like my
skin :).
--------------------------------------------------
Wednesday, 18 September 2019 6.53.29:

So the reading part, related for the User CRUD on the
client side seems to be soon coming to an end.

Next step: pass a Router Control to the User List and
Editor component so that the User/Admin can select an
User by clicking on the corresponding button in the
User-List (which is now in the upper part of the page,
after the text "result for the GetAllUsersReq is:").
--------------------------------------------------