Monday, 23 September 2019 3.08.52:
Introduced the "PostRequestType", phantom type.

Now I have lot of type errors, all over the place.

They either need to be fixed, or we roll back and
remove this phantom type parameter.
--------------------------------------------------
Monday, 23 September 2019 18.20.06:

Phantom type is still in the code.

I used a _ for the type parameter at a lot of places.
Also I have a few ???-s and a few todo-now-s, so,
even though the code compiles now, it will fail
at runtime, most likely.

So the ???-s and todo-now-s have to be fixed, and
after that we need to see how to separate the
"write" and "read" cache / logic in the app, for
which the phantom type parameter was introduced.

Possibly different type classes for the caches, or
some sort of "master cache", or "cache supervisor",
an an extra layer of abstraction above the current
client caching ... we need to see... it's not
clear yet.

A bit of experimentation is ahead on the road.